Move memory functions from lv_uefi_context.c to lv_mem_core_uefi.c
